Centralize beans test configurations in the shared parent
Move DefaultBeanAnalyzerTestConfig (+ CustomDatabaseQualifier) and the
qualifier *Declarations configurations (+ CustomQualifier) out of the
DefaultBeanMetaInfoExtractorTest and QualifiersPersistencePostProcessorTest
subclasses and into AbstractBeansSharedContextTest as nested classes.

The shared parent now owns every test configuration consumed by these
tests and imports only its own nested types, so it never references its
children. The subclasses are reduced to their @Test methods, resolving
the shared fixtures through inheritance. This is a pure refactor: the
cached context composition is unchanged.
